retorno formatado no calculo

+++ src/main/java/com/ficticiusclean/ficticiusclean/controller/VeiculoController.java
@@ -1,16 +1,15 @@
 package com.ficticiusclean.ficticiusclean.controller;
 
+import com.ficticiusclean.ficticiusclean.controller.mapper.VeiculoServiceRequestMapper;
+import com.ficticiusclean.ficticiusclean.controller.mapper.VeiculoServiceResponseMapper;
 import com.ficticiusclean.ficticiusclean.controller.model.response.VeiculoModelRequest;
 import com.ficticiusclean.ficticiusclean.controller.model.response.VeiculoModelResponse;
 import com.ficticiusclean.ficticiusclean.model.VeiculoModel;
 import com.ficticiusclean.ficticiusclean.service.VeiculoService;
-import com.ficticiusclean.ficticiusclean.controller.mapper.VeiculoServiceRequestMapper;
-import com.ficticiusclean.ficticiusclean.controller.mapper.VeiculoServiceResponseMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -29,9 +28,9 @@ public class VeiculoController {
     }
 
     @PostMapping("/calcula")
-    public ResponseEntity<List<VeiculoModelResponse>> calcula(@RequestParam BigDecimal preco,
-                                                              @RequestParam BigDecimal distanciaCidade,
-                                                              @RequestParam BigDecimal distanciaEstrada){
+    public ResponseEntity<List<VeiculoModelResponse>> calcula(@RequestParam Double preco,
+                                                              @RequestParam Double distanciaCidade,
+                                                              @RequestParam Double distanciaEstrada){
         List<VeiculoModelResponse> retorno = service.somaTotal(preco, distanciaCidade, distanciaEstrada);
         return ResponseEntity.ok().body(retorno);
     }
