+++ src/main/java/com/ficticiusclean/ficticiusclean/model/VeiculoModel.java
@@ -20,12 +20,12 @@ public class VeiculoModel implements Serializable {
     private String marca;
     private String modelo;
     private LocalDate dataFabricação;
-    private BigDecimal consumoCombustivelCidade;
-    private BigDecimal consumoCombustivelRodovia;
+    private  Double consumoCombustivelCidade;
+    private  Double consumoCombustivelRodovia;
 
     public VeiculoModel(Long id, String nome, String marca, String modelo,
-                        LocalDate dataFabricação, BigDecimal consumoCombustivelCidade,
-                        BigDecimal consumoCombustivelRodovia) {
+                        LocalDate dataFabricação,  Double consumoCombustivelCidade,
+                        Double consumoCombustivelRodovia) {
         this.id = id;
         this.nome = nome;
         this.marca = marca;
@@ -77,19 +77,19 @@ public class VeiculoModel implements Serializable {
         this.dataFabricação = dataFabricação;
     }
 
-    public BigDecimal getConsumoCombustivelCidade() {
+    public  Double getConsumoCombustivelCidade() {
         return consumoCombustivelCidade;
     }
 
-    public void setConsumoCombustivelCidade(BigDecimal consumoCombustivelCidade) {
+    public void setConsumoCombustivelCidade( Double consumoCombustivelCidade) {
         this.consumoCombustivelCidade = consumoCombustivelCidade;
     }
 
-    public BigDecimal getConsumoCombustivelRodovia() {
+    public  Double getConsumoCombustivelRodovia() {
         return consumoCombustivelRodovia;
     }
 
-    public void setConsumoCombustivelRodovia(BigDecimal consumoCombustivelRodovia) {
+    public void setConsumoCombustivelRodovia( Double consumoCombustivelRodovia) {
         this.consumoCombustivelRodovia = consumoCombustivelRodovia;
     }
 }
